1.6.14

+++ AffiseAttributionLib/Classes/storages/EventsStorageImpl.swift
@@ -1,11 +1,3 @@
-//
-//  EventsStorageImpl.swift
-//  app
-//
-//  Created by Sergey Korney
-//
-
-
 class EventsStorageImpl {
     
     private let logsManager: LogsManager
@@ -110,7 +102,7 @@ extension EventsStorageImpl: EventsStorage {
             
             let events = try directoryContents.filter({ fileURL in
                 let attr = try fileManager.attributesOfItem(atPath: fileURL.path)
-                let creationDate = (attr[.modificationDate] as! Date)
+                guard let creationDate = attr[.modificationDate] as? Date else { return true }
                 if (creationDate.timeIntervalSince1970 < Date().timeIntervalSinceNow - Double(EventsParams.EVENTS_STORE_TIME)) {
                     do {
                         try fileManager.removeItem(atPath: fileURL.path)
